bigserial & bigint => int, domain[]=>text[] (#69)

+++ org_init_scripts/069_personal_use_purchase_item.sql
@@ -1,10 +1,10 @@
 create table if not exists personal_use_purchase_inv_item
 (
     id                       uuid    not null primary key,
-    personal_use_purchase_id bigint  not null,
-    batch_id                 bigint  not null,
-    inventory_id             bigint  not null,
-    unit_id                  bigint  not null,
+    personal_use_purchase_id int  not null,
+    batch_id                 int  not null,
+    inventory_id             int  not null,
+    unit_id                  int  not null,
     unit_conv                float   not null,
     gst_tax_id               text    not null,
     qty                      float   not null,
