+++ org_init_scripts/214_pos_settlement_query.sql
@@ -1,7 +1,7 @@
-create function pos_settlement_breakup_summary(settlement_id bigint)
+create function pos_settlement_breakup_summary(settlement_id int)
     returns table
             (
-                account_id   bigint,
+                account_id   int,
                 account_name text,
                 credit       float,
                 debit        float,
@@ -23,7 +23,7 @@ end;
 $$ language plpgsql immutable
                     security definer;
 --##
-create function pos_settlement_transaction_summary(settlement_id bigint)
+create function pos_settlement_transaction_summary(settlement_id int)
     returns table
             (
                 base_voucher_type text,
